Fixed header icon width
In the main branch of this file, there is CSS specifying that the 120px wide firefox icon in the header should be 100px. Not only is this wrong causing the image to stretch, the user should have to specify the width and height in the <img> as practice. Readers were instructed to do so earlier in the web docs.